Add a timer-based concurrent PID exe search
This modifies the application a bit. Instead of having to press
"Find PID" repeatedly until the EXE is found, now pressing "Start"
triggers a busy loop that keeps looking for the process every
second. If the process is then closed, it goes back to "monitoring"
and sets the PID to zero.